services/python-stream/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
kafka-python==2.0.2
praw==7.6.0
prawcore==2.3.0
requests==2.27.1
update-checker==0.18.0
urllib3==1.26.9
websocket-client==1.3.2
